+++ src/holocron/formula/__init__.py
@@ -1,19 +1,25 @@
 """Safe formula specifications for Holocron."""
 
 from holocron.formula.ast import (
+    CategoricalTerm,
     Formula,
     IdentityTerm,
     LinearSplineTerm,
+    OrderedTerm,
     PolynomialTerm,
     RestrictedCubicSplineTerm,
+    RestrictedInteractionTerm,
     Variable,
 )
 
 __all__ = [
+    "CategoricalTerm",
     "Formula",
     "IdentityTerm",
     "LinearSplineTerm",
+    "OrderedTerm",
     "PolynomialTerm",
     "RestrictedCubicSplineTerm",
+    "RestrictedInteractionTerm",
     "Variable",
 ]